Add is_out_of_date function
This function can be called to determine if a function passed
to jobstamp.run would actually be called again, which may
be useful in the sense that certain peripheral steps can
be skipped.

Internally, this uses the exact same logic as the old run
function, both of which were refactored into the
internal _is_out_of_date function.

bumpversion: minor